+++ oalg-base/CHANGELOG.md
@@ -125,6 +125,7 @@ Empty2 to EntEmpty2, empty2 to fromEmpty2.
 
 ### Added
 - AlgebraicSemiring in the module OAlg.Structure.Algebraic.Definition.
+- mapping cnzMap and cnztMap in the module OAlg.Limes.Exact.ConsZero.
 
 ### Changed
 - eliminated the constraint for HomOriented and HomFibred being an instance of Entity2